refactor: Move SearchController and its response type into rest
SearchController is a @RestController, so it belongs in the rest package
alongside ContentItemController. SearchResultsWithContent is only used by
SearchController (and wraps the raw Elasticsearch SearchResponse, a
web-response shape rather than a domain model), so it moves with it.
SearchControllerTests mirrors the move. contentitem.elasticsearch is now
purely the content-aware indexing concern (indexer + ESRecordTransformers).
