+++ CLAUDE.md
@@ -22,7 +22,7 @@ Most tests are integration tests (tagged `IntegrationTest`, see `TestTypeTags`) 
 
 ## Architecture
 
-**Package layout.** Content-item-agnostic infrastructure lives in top-level packages — `common` (`UuidV7`), `persistence` (`JdbcConfig`, `SchemalessData`), and `elasticsearch` (generic ES infra: `ElasticSearchConfig`, `ElasticSearchIndexCreator`, `IndexController`, the `SearchResponse` (de)serializers). The domain lives under `contentitem`: `ContentItemService`/`ContentItemRepository` at its root, with sub-packages `rest` (controller + exception handler), `exceptions` (the domain exception types), `model` (entity, DTOs, mapper), `transformation` (inbound `ContentItemEntityTransformer`s + `TransformationHandler`), and `elasticsearch` (the content-aware indexer, search controller, and `ESRecordTransformer`s — distinct from the generic top-level `elasticsearch`).
+**Package layout.** Content-item-agnostic infrastructure lives in top-level packages — `common` (`UuidV7`), `persistence` (`JdbcConfig`, `SchemalessData`), and `elasticsearch` (generic ES infra: `ElasticSearchConfig`, `ElasticSearchIndexCreator`, `IndexController`, the `SearchResponse` (de)serializers). The domain lives under `contentitem`: `ContentItemService`/`ContentItemRepository` at its root, with sub-packages `rest` (the `@RestController`s — `ContentItemController` and `SearchController` — the exception handler, and the `SearchResultsWithContent` response type), `exceptions` (the domain exception types), `model` (entity, DTOs, mapper), `transformation` (inbound `ContentItemEntityTransformer`s + `TransformationHandler`), and `elasticsearch` (the content-aware indexer and `ESRecordTransformer`s — distinct from the generic top-level `elasticsearch`).
 
 Ids are **intrinsic** (server-generated UUIDv7, assigned in `ContentItemService`); clients never supply them. The core flow is a dual-write pipeline with two write entry points:
